Move resource content JSON schemas to shared location.
